Add add_entry_2PointO

+++ lib/models/routes_for_client_id_response.dart
@@ -94,7 +94,7 @@ class Hubs {
     int id,
     int hub,
     int sequence,
-    int kms,
+    double kms,
     int tag,
   }) =>
       Hubs(
